# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CreateSnippetRequest:
    properties:
      content:
        description: 100KB max
        maxLength: 100000
        type: string
      label:
        maxLength: 255
        type: string
      shortcut:
        description: Short string without spaces
        maxLength: 50
        type: string
      tags:
        description: Max 20 tags, each max 50 chars
        items:
          type: string
        maxItems: 20
        type: array
    required:
    - content
    - label
    - shortcut
    type: object
...
  models.Snippet:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      label:
        type: string
      shortcut:
        type: string
      tags:
        items:
          type: string
        type: array
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  models.UpdateSnippetRequest:
    properties:
      content:
        type: string
      label:
        type: string
      shortcut:
        type: string
      tags:
        items:
          type: string
        type: array
      userId:
        description: UUID as string
        type: string
    type: object
  models.UpdateUserRequest:
    properties:
      avatarUrl:
        type: string
      email:
        type: string
      fullName:
        type: string
      password:
        description: Will be hashed
        type: string
      username:
        type: string
    type: object
  models.User:
    properties:
      avatarUrl:
        type: string
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        description: UUID as string
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Code snippets management API with authentication
  title: Snippy API
  version: "1.0"
paths:
...
  /snippets:
    get:
      description: Get all snippets belonging to the authenticated user
      parameters:
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Search in label
        in: query
        name: search
        type: string
      - description: Limit results (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user's snippets
      tags:
      - snippets
    post:
      consumes:
      - application/json
      description: Create a new snippet for the authenticated user
      parameters:
      - description: Snippet data
        in: body
        name: snippet
        required: true
        schema:
          $ref: '#/definitions/models.CreateSnippetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Snippet'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a snippet
      tags:
      - snippets
  /snippets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a snippet (owner only)
      parameters:
      - description: Snippet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a snippet
      tags:
      - snippets
    get:
      consumes:
      - application/json
      description: Get a single snippet by its ID
      parameters:
      - description: Snippet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Snippet'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get snippet by ID
      tags:
      - snippets
    put:
      consumes:
      - application/json
      description: Update an existing snippet (owner only)
      parameters:
      - description: Snippet ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: snippet
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSnippetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Snippet'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a snippet
      tags:
      - snippets
...
  /users/profile:
    get:
      description: Get the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the profile of the authenticated user
      parameters:
      - description: Update data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"